Extract common python code into library
